LayoutTests/platform/glib/http/tests/websocket/tests/hybi/handshake-fail-by-mismatch-protocol-header-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/handshake-fail-by-mismatch-protocol-header' failed: Server requested unsupported protocol
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/handshake-fail-by-mismatch-protocol-header' failed: Server requested unsupported protocol
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/handshake-fail-by-mismatch-protocol-header' failed: Server requested unsupported protocol
Test whether WebSocket handshake fails if the server sends mismatched Sec-WebSocket-Protocol header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS closeEvent.wasClean is false
PASS closeEvent.wasClean is false
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
